+++ docs/src/release_notes.rst
@@ -3,6 +3,11 @@
 Release Notes
 #############
 
+2.0.3
+=====
++ Updated ReadArgs decorator to return None when arg doesn't exist.
++ Updated ReadArgs to not log input value.
+
 2.0.2
 =====
 + Updated requirement for stdlib-list to >= 0.6.0 to support Python 3.8.